requirements.txt
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
numpy>=1.20.0
tqdm>=4.65.0
pandas>=1.5.0